Sistema-Analise-Financeira/requirements.txt
Flask==2.3.3
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
plotly>=5.0.0
requests>=2.28.0
python-dotenv>=0.19.0
xlsxwriter>=3.0.0
gunicorn==21.2.0
